+++ changelog.md
@@ -1,9 +1,18 @@
 # Changelog
 
-1.0.2
+1.2.0 (2021-05-01) - Save Compatible
+- Bigger, more obvious spikes
+- Darker purple (easier to see on some light Unknown Skies backgrounds)
+
+1.1.0 (2021-04-20) - Save Compatible
+- Increased usage to 1.8 volatiles per day
+- It's still really good
+- Added a config file to change the cost (`data/config/modSettings.json`)
+
+1.0.2 - Save Compatible
 - Fixed the ability sometimes not getting added (why is the normal neutrino scan sometimes in one list and sometimes in another?)
 
-1.0.1
+1.0.1 - Save Compatible
 - Fixed incompatibility with version checker
 
 1.0